refactor(migration): generalize the flow close header into a step header with back
